Adding a method to generate encounters

+++ src/main/java/br/com/canoza/utils/GenerationUtils.java
@@ -4,7 +4,7 @@ public class GenerationUtils {
     private GenerationUtils() {
     }
 
-    public static int random(final int maxValue) {
+    public static int random(final long maxValue) {
         return (int) (Math.random() * maxValue);
     }
 }
